# docs/api.yaml
openapi: 3.0.0
info:
  title: authentication
  version: 1.0.0
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with the provided registration details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - login
                - password
              properties:
                name:
                  type: string
                  example: John Doe
                login:
                  type: string
                  example: johndoe
                password:
                  type: string
                  format: password
                  example: StrongP@ssw0rd
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully.
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Login already in use.
  /auth/login:
    post:
      summary: Authenticate a user
      description: Logs in an existing user and returns an authentication token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
              properties:
                login:
                  type: string
                  example: johndoe
                password:
                  type: string
                  format: password
                  example: StrongP@ssw0rd
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid login or password.
  /auth/profile:
    get:
      summary: Get the authenticated user's profile
      description: Returns the profile information of the currently authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: John Doe
                  login:
                    type: string
                    example: johndoe
        '401':
          description: Unauthorized - authentication token is missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized access.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT